Scheduling, EWT, Singapore model bugs
git-svn-id: file:///srv/local/data/cleanRepo/trunk@25853 a0517364-e873-4167-b1c8-416daa5a6b91

+++ playgrounds/sergioo/src/main/java/playground/sergioo/eventAnalysisTools2012/gui/DynamicVariableSizeSelectionNetworkPainter.java
@@ -123,7 +123,7 @@ public class DynamicVariableSizeSelectionNetworkPainter extends DynamicNetworkPa
 				Double weight = weights.get(link.getId());
 				if(weight!=null && weight>min) {
 					float proportion = (float)((weight-minLinkWeight)/(maxLinkWeight-minLinkWeight));
-					paintCircle(g2, layersPanel, link.getToNode().getCoord(), proportion*maxSize, JetColor.getJetColor(proportion, 0.5f));
+					paintCircle(g2, layersPanel, link.getToNode().getCoord(), (int)(proportion*maxSize), JetColor.getJetColor(proportion, 0.5f));
 				}
 			}
 			for(Node node:dynamicNetworkPainterManager.getSelectedNodes())
